fetch all products using json server
