+++ react-ecommerce/src/app/store.js
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
-import  counterReducer  from "../features/counter/counterSlice"
+import productReducer from "../features/productList/productSlice"
 
 export const store = configureStore({
   reducer: {
-    counter: counterReducer
+    product: productReducer
   },
 });
