+++ react-ecommerce/src/features/productList/productAPI.js
@@ -1,7 +1,23 @@
 // A mock function to mimic making an async request for data
-export function fetchProductList(amount = 1) {
+export function fetchAllProducts() {
   return new Promise(async (resolve) => {
-    const response = await fetch("http://localhost:8080");
+    const response = await fetch("http://localhost:8080/products");
+    const data = await response.json();
+    resolve({ data });
+  });
+}
+
+export function fetchAllProductsByFilter(filter) {
+  //filter={ "category": "smartphone" };
+  let queryString = "";
+  for (let key in filter) {
+    queryString += `${key}=${filter[key]}&`;
+  }
+
+  return new Promise(async (resolve) => {
+    const response = await fetch(
+      "http://localhost:8080/products?" + queryString
+    );
     const data = await response.json();
     resolve({ data });
   });
